requirements.txt
u8darts[torch]
# darts==0.30.0ß
pandas
numpy
scikit-learn
pyarrow
google-auth
google-cloud-storage
google-cloud-bigquery
google-cloud-aiplatform
google-cloud-bigquery-storage
db-dtypes
joblib==1.4.2
